// DataCenterWebApp/ClientApp/src/app/interfaces/isystemoverview.ts
export interface ISystemOverview {
  LastImportDate: Date;

    ExperimentCount: number;

    HighestSequenceID: number;

    ICDataCenterAddress: string;

    ICDataCenterVersion: string;

    ICDataCenterStatus: string;

    DataCenterWebAppAddress: string;

    DataCenterWebAppVersion: string;

    DataCenterWebAppStatus: string;

    LoggedOnUser: string;

    LoggedOnRole: string;

  LastUpdate: Date;
}